Move CustomOverlayRenderer into separate file
